BuildingList.js: passed in props (data, filterText, selectedUpdate) from App.js the buildings that are listed are now filtered based on the value of filterText When a building is click, the id is stored and returned
Search.js: the value from the textbox is saved and updated as it changes.
The onChange value for the input tag captures teh textbox value.

ViewBuilding.js: the information for the building selected is now shown to the right of the building list.
The information that is displayed is name, code, coordinates.

App.js:
The filterText property of state is set to the value passed into the filterUpdate function.
The selectedUpdate property of state is updated to the id passed into the selectedUpdate function.
Search props filterText and filterUpdate were added.
BuildingList props filterText and selectedUpdate were added.
ViewBuilding props selectedBuilding and data were added.